rename decoding package to decoder
