+++ heifconverter/src/main/java/linc/com/heifconverter/decoder/legacy/iso14496/part12/ItemLocationBox.kt
@@ -1,7 +1,7 @@
 //
 // adhoc patch: org.mp4parser.boxes.iso14496.part12.ItemLocationBox
 //
-package linc.com.heifconverter.decoding.legacy.iso14496.part12
+package linc.com.heifconverter.decoder.legacy.iso14496.part12
 
 import org.mp4parser.support.AbstractFullBox
 import org.mp4parser.tools.IsoTypeReader
