+++ heifconverter/src/main/resources/isoparser-custom.properties
@@ -1,8 +1,8 @@
-iinf=linc.com.heifconverter.decoding.legacy.iso14496.part12.ItemInfoBox
-infe=linc.com.heifconverter.decoding.legacy.iso14496.part12.ItemInfoEntry
-iloc=linc.com.heifconverter.decoding.legacy.iso14496.part12.ItemLocationBox
-iprp=linc.com.heifconverter.decoding.legacy.iso14496.part12.ItemPropertiesBox
-ipco=linc.com.heifconverter.decoding.legacy.iso14496.part12.ItemPropertyContainerBox
-ipma=linc.com.heifconverter.decoding.legacy.iso14496.part12.ItemPropertyAssociation
-pitm=linc.com.heifconverter.decoding.legacy.iso14496.part12.PrimaryItemBox
-ispe=linc.com.heifconverter.decoding.legacy.iso23008.part12.ImageSpatialExtentsBox
+iinf=linc.com.heifconverter.decoder.legacy.iso14496.part12.ItemInfoBox
+infe=linc.com.heifconverter.decoder.legacy.iso14496.part12.ItemInfoEntry
+iloc=linc.com.heifconverter.decoder.legacy.iso14496.part12.ItemLocationBox
+iprp=linc.com.heifconverter.decoder.legacy.iso14496.part12.ItemPropertiesBox
+ipco=linc.com.heifconverter.decoder.legacy.iso14496.part12.ItemPropertyContainerBox
+ipma=linc.com.heifconverter.decoder.legacy.iso14496.part12.ItemPropertyAssociation
+pitm=linc.com.heifconverter.decoder.legacy.iso14496.part12.PrimaryItemBox
+ispe=linc.com.heifconverter.decoder.legacy.iso23008.part12.ImageSpatialExtentsBox
